requirements.txt
# Core dependencies
python-dotenv>=1.0
gradio>=4.0

# Optional: MLflow (for CLI with prompt registration/evaluation)
mlflow==3.10.0

# LLM Providers
groq>=0.4.0              # For Groq API
openai>=1.0             # For OpenAI API

# Optional: Data processing (for MLflow evaluation)
pandas>=2.0
datasets>=2.0
scikit-learn>=1.2
click>=8.0

# Development & testing
pytest>=8.0
